refactor: Move code to internal packages
