+++ pkg/cloudencrypt/gcp.go
@@ -6,6 +6,8 @@ import (
 	kms "cloud.google.com/go/kms/apiv1"
 	"cloud.google.com/go/kms/apiv1/kmspb"
 	"github.com/pkg/errors"
+
+	"github.com/keboola/go-cloud-encrypt/pkg/cloudencrypt/internal/encode"
 )
 
 // GCPEncryptor Implements Encryptor using Google Cloud's Key Management Service.
@@ -27,7 +29,7 @@ func NewGCPEncryptor(ctx context.Context, keyID string) (*GCPEncryptor, error) {
 }
 
 func (encryptor *GCPEncryptor) Encrypt(ctx context.Context, plaintext []byte, metadata Metadata) ([]byte, error) {
-	additionalData, err := encode(metadata)
+	additionalData, err := encode.Encode(metadata)
 	if err != nil {
 		return nil, err
 	}
@@ -47,7 +49,7 @@ func (encryptor *GCPEncryptor) Encrypt(ctx context.Context, plaintext []byte, me
 }
 
 func (encryptor *GCPEncryptor) Decrypt(ctx context.Context, ciphertext []byte, metadata Metadata) ([]byte, error) {
-	additionalData, err := encode(metadata)
+	additionalData, err := encode.Encode(metadata)
 	if err != nil {
 		return nil, err
 	}
