+++ pkg/cloudencrypt/internal/encode/encode_test.go
@@ -1,25 +1,27 @@
-package cloudencrypt
+package encode
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/keboola/go-cloud-encrypt/pkg/cloudencrypt/internal/random"
 )
 
 func TestEncodeDecode(t *testing.T) {
 	t.Parallel()
 
-	secretKey, err := generateSecretKey()
+	secretKey, err := random.SecretKey()
 	assert.NoError(t, err)
 
 	data := make(map[string][]byte)
 	data["test"] = secretKey
 
-	encoded, err := encode(data)
+	encoded, err := Encode(data)
 	assert.NoError(t, err)
 	assert.NotNil(t, encoded)
 
-	decoded, err := decode[map[string][]byte](encoded)
+	decoded, err := Decode[map[string][]byte](encoded)
 	assert.NoError(t, err)
 	assert.NotNil(t, decoded)
